test: externalize suite env generatoin
